Sepaate the delete functionality
