+++ components/expenses/DeleteExpenseButton.js
@@ -1,48 +1,15 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { gql } from '@apollo/client';
-import { Mutation } from '@apollo/client/react/components';
 import { Trash2 as IconTrash } from '@styled-icons/feather/Trash2';
-import { FormattedMessage, useIntl } from 'react-intl';
+import { FormattedMessage } from 'react-intl';
 
-import { i18nGraphqlException } from '../../lib/errors';
-import { API_V2_CONTEXT } from '../../lib/graphql/helpers';
-
-import ConfirmationModal from '../ConfirmationModal';
 import StyledButton from '../StyledButton';
 import { Span } from '../Text';
-import { TOAST_TYPE, useToasts } from '../ToastProvider';
 
-export const deleteExpenseMutation = gql`
-  mutation DeleteExpense($id: String!) {
-    deleteExpense(expense: { id: $id }) {
-      id
-    }
-  }
-`;
-
-export const removeExpenseFromCache = (cache, { data: { deleteExpense } }) => {
-  cache.modify({
-    fields: {
-      expenses(existingExpenses, { readField }) {
-        if (!existingExpenses?.nodes) {
-          return existingExpenses;
-        } else {
-          return {
-            ...existingExpenses,
-            totalCount: existingExpenses.totalCount - 1,
-            nodes: existingExpenses.nodes.filter(expense => deleteExpense.id !== readField('id', expense)),
-          };
-        }
-      },
-    },
-  });
-};
+import ExpenseConfirmDeletion from './ExpenseConfirmDeletionModal';
 
 const DeleteExpenseButton = ({ expense, onDelete, buttonProps, isDisabled, onModalToggle }) => {
   const [hasDeleteConfirm, setDeleteConfirm] = React.useState(false);
-  const { addToast } = useToasts();
-  const intl = useIntl();
   const showDeleteConfirm = isOpen => {
     setDeleteConfirm(isOpen);
     onModalToggle?.(isOpen);
@@ -63,33 +30,11 @@ const DeleteExpenseButton = ({ expense, onDelete, buttonProps, isDisabled, onMod
         </Span>
       </StyledButton>
       {hasDeleteConfirm && (
-        <Mutation mutation={deleteExpenseMutation} context={API_V2_CONTEXT} update={removeExpenseFromCache}>
-          {deleteExpense => (
-            <ConfirmationModal
-              isDanger
-              type="delete"
-              onClose={() => showDeleteConfirm(false)}
-              header={<FormattedMessage id="actions.delete" defaultMessage="Delete" />}
-              continueHandler={async () => {
-                try {
-                  await deleteExpense({ variables: { id: expense.id } });
-                } catch (e) {
-                  addToast({ type: TOAST_TYPE.ERROR, message: i18nGraphqlException(intl, e) });
-                }
-
-                if (onDelete) {
-                  await onDelete(expense);
-                }
-                showDeleteConfirm(false);
-              }}
-            >
-              <FormattedMessage
-                id="Expense.DeleteDetails"
-                defaultMessage="This will permanently delete the expense and all attachments and comments."
-              />
-            </ConfirmationModal>
-          )}
-        </Mutation>
+        <ExpenseConfirmDeletion
+          onDelete={onDelete}
+          expense={expense}
+          showDeleteConfirmMoreActions={showDeleteConfirm}
+        />
       )}
     </React.Fragment>
   );
